move Category and Location to entity package
