--- src/main/java/edu/tbank/hw5/dto/Category.java
@@ -1,16 +1,0 @@
-package edu.tbank.hw5.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class Category implements Serializable {
-    private Long id;
-    private String slug;
-    private String name;
-}
