+++ src/main/java/edu/tbank/hw5/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
 package edu.tbank.hw5.repository;
 
-import edu.tbank.hw5.dto.Category;
+import edu.tbank.hw5.entity.Category;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
