+++ src/main/java/edu/tbank/hw5/service/impl/CategoryServiceImpl.java
@@ -1,6 +1,6 @@
 package edu.tbank.hw5.service.impl;
 
-import edu.tbank.hw5.dto.Category;
+import edu.tbank.hw5.entity.Category;
 import edu.tbank.hw5.repository.CategoryRepository;
 import edu.tbank.hw5.service.CategoryService;
 import lombok.RequiredArgsConstructor;
