+++ src/main/java/edu/tbank/hw5/service/impl/LocationServiceImpl.java
@@ -1,6 +1,6 @@
 package edu.tbank.hw5.service.impl;
 
-import edu.tbank.hw5.dto.Location;
+import edu.tbank.hw5.entity.Location;
 import edu.tbank.hw5.repository.LocationRepository;
 import edu.tbank.hw5.service.LocationService;
 import lombok.RequiredArgsConstructor;
